libudev: hide definition of struct udev_device
